Incorporate label ID into Vuex state
Move label ID retrieval into Vuex state so that the input data ID and
label IDs are updated atomically.
This is to prevent the previous issue where input data ID would be updated
in the parent component, and the label ID would be reactively updated in
the child component, and sometimes there would be misalignment of the two,
due to asynchronous functions being run.