Correct for padding offset when generating binary images
This is a temporary solution to fix the accidental canvas padding offset
in the polygon data that is stored in the database.
To fix the problem properly, the offset must be made to the data in the DB
and the addition and subtracion of the offsets on the front end must
be switched around.